Finish Code refactoring
